Implement more interfaces and start converting spigot to use the api

+++ Spigot/src/main/java/me/darkeyedragon/randomtp/config/section/ConfigQueue.java
@@ -21,7 +21,7 @@ public class ConfigQueue implements SectionQueue {
         return size;
     }
 
-    public int getInitDelay() {
+    public long getInitDelay() {
         return initDelay;
     }
 }
